modularisering_0.3
startet med endring av filstruktur og pom inheritance, WIP
