+++ careconnect-security/src/test/java/com/careconnect/security/SecurityConfigTest.java
@@ -1,4 +1,4 @@
-package com.careconnect.careconnect;
+package com.careconnect.security;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -17,7 +17,7 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.redirectedUrl;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.careconnect.careconnect.config.SecurityConfig;
+import com.careconnect.security.config.SecurityConfig;
 import com.careconnect.careconnect.controller.AdminController;
 import com.careconnect.careconnect.controller.LoginController;
 import com.careconnect.careconnect.controller.UserController;
