+++ pom.xml
@@ -1,0 +1,48 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>com.careconnect</groupId>
+    <artifactId>careconnect-parent</artifactId>
+    <version>1.0-SNAPSHOT</version>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>careconnect</module>
+        <module>careconnect-security</module>
+        <!-- referer til flere moduler her -->
+    </modules>
+
+    <dependencyManagement>
+        <dependencies>
+        <dependency>
+        <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
+            <version>5.9.0</version>
+            <scope>test</scope>
+            </dependency>
+            <!-- definer felles dependencies her, sammenlign pom.xml til forskjellige moduler mot slutten av prosjektet -->
+            <!-- må referere til denne pom.xml filen i barne pom.xmlene -->
+            <dependency>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-starter</artifactId>
+                <version>3.0.0</version>
+            </dependency>
+            <dependency>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-starter-security</artifactId>
+                <version>3.0.0</version>
+            </dependency>
+            
+        </dependencies>
+    </dependencyManagement>
+
+    <build>
+        <pluginManagement>
+            <plugins>
+                <!-- definer felles plugins her, sammenlign pom.xml til forskjellige moduler mot slutten av prosjektet -->
+                <!-- må referere til denne pom.xml filen i barne pom.xmlene -->
+            </plugins>
+        </pluginManagement>
+    </build>
+</project>
